Entirely restructure analyzer
- Create subdirectory for HTML templates
- Split most functions out of loganalyzer.py into checks/
- Convert log file regex checks to reusable functions
- Added requirements.txt to document dependencies
- Added setup.cfg to configure rules for pycodestyle
- All code formatted with pycodestyle
- Improved Readme documentation
- Added super basic GitHub Actions workflow

This commit does not change any core functionality, just shuffles code
around and simplifies the readability.
