+++ README.md
@@ -1,14 +1,42 @@
 # loganalyzer
 Automatic analyzer for logs generated by OBS Studio
 
+## Installation
 
-## Webserver
-Start `simplehttp.py` and there will be a website available at (http://localhost:8080)
+### Basic Requirements
 
-## Terminal edition
+* Python 3.x
+
+### Dependencies
+
+```bash
+pip install -r requirements.txt
+```
+
+`pycodestyle` is used for code formatting.
+
+## Usage
+
+### Web Server
+
+```bash
+$ ./simplehttp.py
+Serving on "localhost" with port "8080"
+```
+
+Once launched, available at http://localhost:8080
+
+Includes a basic JSON API
+
+```plain
+http://localhost:8080/?format=json&url=
+```
+
+### Terminal
+
 Start `loganalyzer.py` in your favourite terminal.
 
-```
+```bash
 $ ./loganalyzer.py -h
 usage: loganalyzer.py [-h] --url URL
 
@@ -18,12 +46,13 @@ optional arguments:
 ```
 
 The log specifide by `URL` will be automatically analyzed. Output looks like:
-```
-Critical: AUTOCONFIG WIZARD, INSUFFICIENT HARDWARE, MP4 RECORDING, RENDERING LAG, CPU OVERLOAD, 
+
+```plain
+Critical: AUTOCONFIG WIZARD, INSUFFICIENT HARDWARE, MP4 RECORDING, RENDERING LAG, CPU OVERLOAD,
 Warning:
-Info:     NOT ADMIN, LOW STREAM BANDWIDTH, NO SCENES/SOURCES, 
+Info:     NOT ADMIN, LOW STREAM BANDWIDTH, NO SCENES/SOURCES,
 --------------------------------------
- 
+
 Details
 
 Critical:
@@ -39,5 +68,3 @@ NOT ADMIN
     OBS is not running as administrator. This can lead to obs not being able to
     	....
 ```
-
-
