+++ check_pep8.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
 set -e
 
-# ignore long lines (prefer tesult text in sinle line)
-pycodestyle --ignore=E501 .
+# ignore long lines (prefer result text in sinle line)
+pycodestyle --ignore=E501,E722 .
