--- detail.html
@@ -1,6 +1,0 @@
-<div id="{anchor}" class="card text-white border-{sev} mb-3">
-	<h5 class="card-header"><span class="badge badge-pill badge-{sev} align-text-bottom mr-2">{severity}</span> {title}</h5>
-	<div class="card-body">
-		<p class="card-text">{text}</p>
-	</div>
-</div>
