+++ setup.cfg
@@ -1,0 +1,2 @@
+[pycodestyle]
+ignore = E501, E722
