+++ templates/index.html
@@ -1,18 +1,21 @@
 <!DOCTYPE HTML>
 <html>
+
 <head>
-    <meta charset="utf-8">
-    <title>OBS Log Analyzer</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="https://bootswatch.com/4/darkly/bootstrap.min.css">
+	<meta charset="utf-8">
+	<title>OBS Log Analyzer</title>
+	<meta name="viewport" content="width=device-width, initial-scale=1">
+	<link rel="stylesheet" href="https://bootswatch.com/4/darkly/bootstrap.min.css">
 </head>
+
 <body>
-    <div class="container">
+	<div class="container">
 		<div class="page-header my-5" id="banner">
 			<div class="row">
 				<div class="col-lg-8 col-md-7 col-sm-6">
 					<h1>OBS Log Analyzer</h1>
-					<p class="lead">Please <a href="https://github.com/obsproject/loganalyzer/issues/new">report</a> all crashes together with the log that caused a crash.</p>
+					<p class="lead">Please <a href="https://github.com/obsproject/loganalyzer/issues/new">report</a> all
+						crashes together with the log that caused a crash.</p>
 				</div>
 			</div>
 		</div>
@@ -31,7 +34,8 @@
 									<div class="input-group-prepend">
 										<span class="input-group-text">Log URL input</span>
 									</div>
-									<input class="form-control" id="logUrlInput" aria-describedby="urlHelp" placeholder="Paste log URL here" value="{ph}" name="url" type="text">
+									<input class="form-control" id="logUrlInput" aria-describedby="urlHelp"
+										placeholder="Paste log URL here" value="{ph}" name="url" type="text">
 									<div class="input-group-append">
 										<button type="submit" class="btn btn-primary">Analyze</button>
 									</div>
@@ -55,7 +59,7 @@
 					<h5 class="card-header">Critical</h5>
 					<div class="card-body">
 						{summary_critical}
-					 </div>
+					</div>
 				</div>
 			</div>
 			<div class="col">
@@ -63,7 +67,7 @@
 					<h5 class="card-header">Warning</h5>
 					<div class="card-body">
 						{summary_warning}
-					 </div>
+					</div>
 				</div>
 			</div>
 			<div class="col">
@@ -71,11 +75,12 @@
 					<h5 class="card-header">Info</h5>
 					<div class="card-body">
 						{summary_info}
-					 </div>
+					</div>
 				</div>
 			</div>
 		</div>
 		{details}
-    </div>
+	</div>
 </body>
+
 </html>
